♻️ refactor: 파일 경로 수정(#390)
page들은 component폴더가 아닌 mypage 폴더 바로 밑에 위치하도록 수정
